requirements.txt
apparmor==2.12
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
decorator==4.4.1
distro-info===0.18ubuntu0.18.04.1
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
LibAppArmor==2.12
netifaces==0.10.4
oauthlib==3.1.0
py==1.8.0
pycrypto==2.6.1
pygobject==3.26.1
pymongo==3.9.0
python-apt==1.6.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-oauthlib==1.2.0
retry==0.9.2
SecretStorage==2.3.1
six==1.11.0
ssh-import-id==5.7
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
